Move `Shape` to own file
